[MERGE]8.like-bookmarks: Completed. Added like/unlike logic and views.

+++ app/controllers/likes_controller.rb
@@ -1,29 +1,36 @@
 #
 class LikesController < ApplicationController
   def index
+    authorize Like
     @likes = Like.all
   end
 
   def create
     @bookmark = Bookmark.find(params[:bookmark_id])
     like = current_user.likes.build(bookmark: @bookmark)
+    authorize like
 
     if like.save
-      redirect_to @bookmark, notice: 'success'
+      # Response: Redisplay bookmark view with new like state, msgs.
+      redirect_to :back, notice: 'Bookmark was Liked successfully.'
     else
-      flash.now(:alert)
-      redirect_to @bookmark
+      # Response: Redisplay bookmarks view with err msgs.
+      redirect_to :back, alert: 'Error creating like. Please try again.'
     end
   end
 
   def destroy
     @bookmark = Bookmark.find(params[:bookmark_id])
     # Find the current user's like with the ID in the params
+    like = Like.find(params[:id])
+    authorize like
 
     if like.destroy
-      # Flash success and redirect to @bookmark
+      # Response: Redisplay bookmark view with new like state, msgs.
+      redirect_to :back, notice: 'Bookmark was UnLiked successfully.'
     else
-      # Flash error and redirect to @bookmark
+      # Response: Redisplay bookmarks view with err msgs.
+      redirect_to :back, alert: 'Error destroying like. Please try again.'
     end
   end
 end
